+++ lib/archetype/sass_extensions/functions/styleguide/grammar.rb
@@ -13,8 +13,7 @@ module Archetype::SassExtensions::Styleguide
   # - {Array} an array containing the identifer, modifiers, and a token
   #
   def grammar(sentence, theme = nil, state = nil)
-    debug = Compass.configuration.styleguide_debug
-    helpers.debug "[archetype:styleguide] converting `#{sentence}` to grammar" if debug
+    _styleguide_debug "converting `#{sentence}` to grammar", :grammar
     theme = get_theme(theme)
     components = theme[:components]
     # get a list of valid ids
@@ -32,10 +31,12 @@ module Archetype::SassExtensions::Styleguide
     # maybe in the case where we're looking for strict keys on the lookup?
     modifiers = modifiers.uniq
     token = memoizer.tokenize(theme[:name], extensions, id, modifiers, state)
-    if debug
-      helpers.debug "[archetype:styleguide] the computed grammar is..."
-      helpers.debug "  identifier: #{id}"
-      helpers.debug "  modifiers: #{modifiers}"
+    if id
+      _styleguide_debug "the computed grammar is...", :grammar
+      _styleguide_debug "  identifier: #{id}", :grammar
+      unless modifiers.empty?
+        _styleguide_debug "  modifiers: #{modifiers.join(', ')}", :grammar
+      end
     end
     return id, modifiers, token
   end